LayoutTests/media/track/track-cue-rendering-expected.txt

Test that default positioned TextTrack's cues are rendered correctly.
EVENT(canplaythrough)
EVENT(seeked)
EXPECTED (video.currentTime == '0.5') OK
EXPECTED (testTrack.track.activeCues.length == '1') OK
EXPECTED (testTrack.track.activeCues[0].text == 'Lorem') OK
EXPECTED (testCueDisplayBox.innerText == 'Lorem') OK
EXPECTED (2 * testCueDisplayBox.offsetLeft == video.videoWidth - testCueDisplayBox.offsetWidth == 'true') OK

RUN(video.currentTime = 1.5)
EVENT(seeked)
EXPECTED (video.currentTime == '1.5') OK
EXPECTED (testTrack.track.activeCues.length == '1') OK
EXPECTED (testTrack.track.activeCues[0].text == 'ipsum') OK
EXPECTED (testCueDisplayBox.innerText == 'ipsum') OK
EXPECTED (2 * testCueDisplayBox.offsetLeft == video.videoWidth - testCueDisplayBox.offsetWidth == 'true') OK

RUN(video.currentTime = 2.5)
EVENT(seeked)
EXPECTED (video.currentTime == '2.5') OK
EXPECTED (testTrack.track.activeCues.length == '1') OK
EXPECTED (testTrack.track.activeCues[0].text == 'dolor') OK
EXPECTED (testCueDisplayBox.innerText == 'dolor') OK
EXPECTED (2 * testCueDisplayBox.offsetLeft == video.videoWidth - testCueDisplayBox.offsetWidth == 'true') OK

RUN(video.currentTime = 3.5)
EVENT(seeked)
EXPECTED (video.currentTime == '3.5') OK
EXPECTED (testTrack.track.activeCues.length == '1') OK
EXPECTED (testTrack.track.activeCues[0].text == 'sit') OK
EXPECTED (testCueDisplayBox.innerText == 'sit') OK
EXPECTED (2 * testCueDisplayBox.offsetLeft == video.videoWidth - testCueDisplayBox.offsetWidth == 'true') OK

Test the cue display colors and font.
RUN(video.width = 320)
RUN(video.height = 240)
EXPECTED (getComputedStyle(textTrackDisplayElement(video)).fontSize == '12px') OK

RUN(video.width = 640)
RUN(video.height = 480)
EXPECTED (getComputedStyle(textTrackDisplayElement(video)).fontSize == '24px') OK

RUN(video.width = 1280)
RUN(video.height = 960)
EXPECTED (getComputedStyle(textTrackDisplayElement(video)).fontSize == '48px') OK

RUN(video.width = 2560)
RUN(video.height = 1440)
EXPECTED (getComputedStyle(textTrackDisplayElement(video)).fontSize == '72px') OK

EXPECTED (getComputedStyle(textTrackDisplayElement(video)).fontFamily == 'sans-serif') OK
EXPECTED (getComputedStyle(textTrackDisplayElement(video)).color == 'rgb(255, 255, 255)') OK
EXPECTED (getComputedStyle(textTrackDisplayElement(video, 'display').firstChild).backgroundColor == 'rgba(0, 0, 0, 0.796875)'), OBSERVED 'rgba(0, 0, 0, 0.8)' FAIL
END OF TEST
